fix: Fix error fd 3 is not a socket on macos
When running on macos, if you have any command that opens file descriptors before systemfd, the following error would happen

```
fd 3 is not a socket
```

One example is `make -j 2 target` as it opens files for its own usage.

This fixes it by creating a variable pointing to the first fd created so listenfd can start from there, instead of the fixed value of 3.