Add configuration for socket listen queue size
Systemd allows setting the backlog for socket services with the
'Backlog' parameter. This helps to support services where the connection
rate is high.

systemfd uses a backlog length of 1, and provides no way to configure
this. Although this utility is intended primarily for testing, there may
still be cases where setting this is useful (for instance when running
performance tests). To support this, this change adds a 'backlog'
parameter, with short code 'b'. The default value for this parameter is
1, which maintains the existing behaviour.

In cases where multiple listening sockets are specified, the same
backlog value is applied to every socket.